types: Use commonjs-style default export
Because the code uses a commonjs-style `module.exports = function ...`
while the types use an ESM-style `export default ...`, the type of the
default export is not intrepreted correctly when using TypeScript 4.7
with `"module": "node16"` (which is currently in RC). Using the same
style of default export in both places works though!