types: Satisfy tests, move SnakeCaseKeys to ns
Interestingly, I found that I could correctly import the type
SnakeCaseKeys without issue in a trivial TS project set to import from
this branch of snakecase-keys, both before and after this change. But
tsd was complaining on the last commit and it is satisfied now, with no
changes to the test itself.